requirement.txt
django>=4.1.0<4.2.5 
djangorestframework
pyyaml
requests
django-cors-headers